Fixed a case where scanner would not delete a series if another series with same name but different format was added in that same scan.
