+++ API.Tests/Helpers/ParserInfoFactory.cs
@@ -1,6 +1,10 @@
-﻿using System.IO;
+﻿using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using API.Entities.Enums;
 using API.Parser;
+using API.Services.Tasks.Scanner;
 
 namespace API.Tests.Helpers
 {
@@ -20,6 +24,50 @@ namespace API.Tests.Helpers
                 Series = series,
                 Volumes = volumes
             };
+        }
+
+        public static void AddToParsedInfo(IDictionary<ParsedSeries, List<ParserInfo>> collectedSeries, ParserInfo info)
+        {
+            var existingKey = collectedSeries.Keys.FirstOrDefault(ps =>
+                ps.Format == info.Format && ps.NormalizedName == API.Parser.Parser.Normalize(info.Series));
+            existingKey ??= new ParsedSeries()
+            {
+                Format = info.Format,
+                Name = info.Series,
+                NormalizedName = API.Parser.Parser.Normalize(info.Series)
+            };
+            if (collectedSeries.GetType() == typeof(ConcurrentDictionary<,>))
+            {
+                ((ConcurrentDictionary<ParsedSeries, List<ParserInfo>>) collectedSeries).AddOrUpdate(existingKey, new List<ParserInfo>() {info}, (_, oldValue) =>
+                {
+                    oldValue ??= new List<ParserInfo>();
+                    if (!oldValue.Contains(info))
+                    {
+                        oldValue.Add(info);
+                    }
+
+                    return oldValue;
+                });
+            }
+            else
+            {
+                if (!collectedSeries.ContainsKey(existingKey))
+                {
+                    collectedSeries.Add(existingKey, new List<ParserInfo>() {info});
+                }
+                else
+                {
+                    var list = collectedSeries[existingKey];
+                    if (!list.Contains(info))
+                    {
+                        list.Add(info);
+                    }
+
+                    collectedSeries[existingKey] = list;
+                }
+
+            }
+
         }
     }
 }
